pyproject.toml
[project]
name = "api-cost"
version = "0.1.0"
description = "LLM API Cost Calculator - 다양한 LLM API의 사용 비용을 계산하고 비교하는 도구"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "dexelop", email = "your.email@example.com" }
]
keywords = ["llm", "api", "cost", "calculator", "openai", "claude", "gemini"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "streamlit>=1.30.0",
    "tiktoken>=0.5.0",
    "PyPDF2>=3.0.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "Pillow>=10.0.0",
    "pandas>=2.0.0",
    "plotly>=5.0.0",
    "PyYAML>=6.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "aiohttp>=3.9.0",
    "chardet>=5.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"
